fix bug with tied chords in repeated music sections
Tokens in the source file can be linked multiple times from the PDF,
if they are contained by a

    myMusic = { ... }

declaration which is then invoked multiple times.

We make the first value in the tuple the coordinate of the left side
of the box, in order that the tuples notePositionsInPage will be
sorted primarily left to right in the order they appear in the PDF,
and secondarily by the order they appear in the source file.