[Issue #39] Ensure interim videos are generated with right format.
If a title page is requested, ly2video creates 2 temporary videos and
then joins them. To avoid lossy changes, the format of these videos
should be the same as that of the requested output file. This should
be as simple as replacing the filename extension of the interim video
with that of the specified output file.