refactor: new 'ScoreImage' class managing score image cropping and travelling
Some VideoFrameWriter methods are now moved to this class.
The getCropTopAndBottomMarginSizes() method is replaced by the topCroppable() and
bottomCroppable() methods which respectivly return the number of pixels the image can
be cropped on top an bottom.
An instance is now used by VideoFrameWriter.

Some unused code has been removed in VideoFrameWriter (write and secsElapsedForTempoChanges)