move ly library to ly2video.ly
ly2video currently uses an old version of the ly module, stolen from
an ancient version of Frescobaldi and imported into the ly2video
source tree under ly/.  Move this under the ly2video namespace, to
prevent clashes with newer installations of the ly module.

Partially fixes #56, but we still need to update ly2video to use a
modern version of the ly module, at which point, we can delete this
imported/stolen version and just get it via pip-requires.txt instead.